Reworked wic installation into file and exec resources
Previously, wic would fail to install on 2003, because it was using a
puppet URI as the package source, which is not supported.

This commit switches it to a file & exec resource. I was unable to use a
package resource because WIC is installed as an update, not as a full
product, so puppet can't detect that it's installed.

This commit is dependent upon the `systemroot` fact.